pyproject.toml
[project]
name = "mouse-wheel-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyobjc-framework-quartz>=11.1",
]
